+++ apps/meteor/app/cloud/server/functions/buildRegistrationData.ts
@@ -1,4 +1,5 @@
-import { Statistics, Users } from '@rocket.chat/models';
+import { LivechatRooms, Statistics, Users } from '@rocket.chat/models';
+import moment from 'moment';
 
 import { settings } from '../../../settings/server';
 import { statistics } from '../../../statistics/server';
@@ -61,6 +62,7 @@ export async function buildWorkspaceRegistrationData<T extends string | undefine
 
 	const { organizationType, industry, size: orgSize, country, language, serverType: workspaceType, registerServer } = stats.wizard;
 	const seats = await Users.getActiveLocalUserCount();
+	const [macs] = await LivechatRooms.getMACStatisticsForPeriod(moment.utc().format('YYYY-MM'));
 
 	return {
 		uniqueId: stats.uniqueId,
@@ -90,7 +92,7 @@ export async function buildWorkspaceRegistrationData<T extends string | undefine
 		setupComplete: setupWizardState === 'completed',
 		connectionDisable: !registerServer,
 		npsEnabled,
-		MAC: stats.omnichannelContactsBySource?.contactsCount ?? 0,
+		MAC: macs?.contactsCount ?? 0,
 		// activeContactsBillingMonth: stats.omnichannelContactsBySource.contactsCount,
 		// activeContactsYesterday: stats.uniqueContactsOfYesterday.contactsCount,
 		statsToken: stats.statsToken,