+++ apps/meteor/app/message-pin/server/pinMessage.ts
@@ -1,4 +1,4 @@
-import { Message } from '@rocket.chat/core-services';
+import { Message, api } from '@rocket.chat/core-services';
 import { isQuoteAttachment, isRegisterUser } from '@rocket.chat/core-typings';
 import type { IMessage, MessageAttachment, MessageQuoteAttachment } from '@rocket.chat/core-typings';
 import { Messages, Rooms, Subscriptions, Users, ReadReceipts } from '@rocket.chat/models';
@@ -9,6 +9,7 @@ import { Meteor } from 'meteor/meteor';
 import { Apps, AppEvents } from '../../../ee/server/apps/orchestrator';
 import { callbacks } from '../../../lib/callbacks';
 import { isTruthy } from '../../../lib/isTruthy';
+import { broadcastMessageSentEvent } from '../../../server/modules/watchers/lib/messages';
 import { canAccessRoomAsync, roomAccessAttributes } from '../../authorization/server';
 import { hasPermissionAsync } from '../../authorization/server/functions/hasPermission';
 import { isTheLastMessage } from '../../lib/server/functions/isTheLastMessage';
@@ -108,9 +109,9 @@ Meteor.methods<ServerMethods>({
 			username: me.username,
 		};
 
-		originalMessage = await Message.beforeSave({ message: originalMessage, room, user: me });
-
 		originalMessage = await callbacks.run('beforeSaveMessage', originalMessage);
+
+		originalMessage = await Message.beforeSave({ message: originalMessage, room, user: me });
 
 		await Messages.setPinnedByIdAndUserId(originalMessage._id, originalMessage.pinnedBy, originalMessage.pinned);
 		if (settings.get('Message_Read_Receipt_Store_Users')) {
@@ -211,9 +212,9 @@ Meteor.methods<ServerMethods>({
 			throw new Meteor.Error('not-authorized', 'Not Authorized', { method: 'unpinMessage' });
 		}
 
-		originalMessage = await Message.beforeSave({ message: originalMessage, room, user: me });
-
 		originalMessage = await callbacks.run('beforeSaveMessage', originalMessage);
+
+		originalMessage = await Message.beforeSave({ message: originalMessage, room, user: me });
 
 		if (isTheLastMessage(room, message)) {
 			await Rooms.setLastMessagePinned(room._id, originalMessage.pinnedBy, originalMessage.pinned);
@@ -226,6 +227,10 @@ Meteor.methods<ServerMethods>({
 		if (settings.get('Message_Read_Receipt_Store_Users')) {
 			await ReadReceipts.setPinnedByMessageId(originalMessage._id, originalMessage.pinned);
 		}
+		void broadcastMessageSentEvent({
+			id: message._id,
+			broadcastCallback: (message) => api.broadcast('message.sent', message),
+		});
 
 		return true;
 	},