+++ apps/meteor/client/views/omnichannel/additionalForms.tsx
@@ -1,7 +1,7 @@
 import BusinessHoursMultipleContainer from '../../../ee/client/omnichannel/additionalForms/BusinessHoursMultipleContainer';
 import ContactManager from '../../../ee/client/omnichannel/additionalForms/ContactManager';
 import CurrentChatTags from '../../../ee/client/omnichannel/additionalForms/CurrentChatTags';
-import CustomFieldsAdditionalFormContainer from '../../../ee/client/omnichannel/additionalForms/CustomFieldsAdditionalFormContainer';
+import CustomFieldsAdditionalForm from '../../../ee/client/omnichannel/additionalForms/CustomFieldsAdditionalForm';
 import DepartmentBusinessHours from '../../../ee/client/omnichannel/additionalForms/DepartmentBusinessHours';
 import DepartmentForwarding from '../../../ee/client/omnichannel/additionalForms/DepartmentForwarding';
 import EeNumberInput from '../../../ee/client/omnichannel/additionalForms/EeNumberInput';
@@ -13,7 +13,7 @@ import PrioritiesSelect from '../../../ee/client/omnichannel/additionalForms/Pri
 import SlaPoliciesSelect from '../../../ee/client/omnichannel/additionalForms/SlaPoliciesSelect';
 
 export {
-	CustomFieldsAdditionalFormContainer,
+	CustomFieldsAdditionalForm,
 	MaxChatsPerAgentContainer,
 	MaxChatsPerAgentDisplay,
 	EeNumberInput,