+++ apps/meteor/tests/end-to-end/api/livechat/10-departments.ts
@@ -344,8 +344,15 @@ import { IS_EE } from '../../../e2e/config/constants';
 	});
 
 	describe('PUT livechat/departments/:_id', () => {
+		let department: ILivechatDepartment;
+		before(async () => {
+			department = await createDepartment();
+		});
+		after(async () => {
+			await deleteDepartment(department._id);
+		});
+
 		it('should return an error if fallbackForwardDepartment points to same department', async () => {
-			const department = await createDepartment();
 			await request
 				.put(api(`livechat/department/${department._id}`))
 				.set(credentials)
@@ -361,7 +368,58 @@ import { IS_EE } from '../../../e2e/config/constants';
 				})
 				.expect('Content-Type', 'application/json')
 				.expect(400);
-			await deleteDepartment(department._id);
+		});
+		it('should fail if `agents` param is not an array', async () => {
+			await request
+				.put(api(`livechat/department/${department._id}`))
+				.set(credentials)
+				.send({
+					department: {
+						name: faker.hacker.adjective(),
+						enabled: true,
+						showOnOfflineForm: true,
+						showOnRegistration: true,
+						email: faker.internet.email(),
+					},
+					agents: 'not an array',
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(400);
+		});
+		it('should throw an error if user has permission to add agents and agents array has invalid format', async () => {
+			await updatePermission('add-livechat-department-agents', ['admin']);
+			await request
+				.put(api(`livechat/department/${department._id}`))
+				.set(credentials)
+				.send({
+					department: {
+						name: faker.hacker.adjective(),
+						enabled: true,
+						showOnOfflineForm: true,
+						showOnRegistration: true,
+						email: faker.internet.email(),
+					},
+					agents: [{ notAValidKey: 'string' }],
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(400);
+		});
+		it('should throw an error if user has permission to add agents and agents array has invalid internal format', async () => {
+			await request
+				.put(api(`livechat/department/${department._id}`))
+				.set(credentials)
+				.send({
+					department: {
+						name: faker.hacker.adjective(),
+						enabled: true,
+						showOnOfflineForm: true,
+						showOnRegistration: true,
+						email: faker.internet.email(),
+					},
+					agents: [{ upsert: [{ notAValidKey: 'string' }] }],
+				})
+				.expect('Content-Type', 'application/json')
+				.expect(400);
 		});
 	});