+++ packages/ui-client/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @rocket.chat/ui-client
 
+## 3.0.0-rc.7
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.0-rc.7
+
+## 3.0.0-rc.6
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.0-rc.6
+
+## 3.0.0-rc.5
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.0-rc.5
+
+## 3.0.0-rc.4
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.0-rc.4
+
+## 3.0.0-rc.3
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.0-rc.3
+
 ## 3.0.0-rc.2
 
 ### Patch Changes
@@ -18,6 +48,24 @@
 
 - Updated dependencies [7da1edf866]
   - @rocket.chat/ui-contexts@3.0.0-rc.0
+
+## 2.0.8
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@2.0.8
+
+## 2.0.7
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@2.0.7
+
+## 2.0.6
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@2.0.6
 
 ## 2.0.5